requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
aiohttp_socks>=0.7.1
fastmcp>=2.9.2
pytest>=6.0.0
python-dotenv>=0.19.0
setuptools>=65.0.0 